feat: Refactor JAX backend to import JaxSweeper from the correct module and add JaxSweeper class implementation
